統合課題，心拍が安定しない

+++ workB/b3_pulse_measurement/myPulse.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 
-uint16_t top();  //ピークを検出する
+uint32_t top();  //ピークを検出する
 uint8_t rate();  //脈拍を測定し返す
 
 #endif
